Move DatabaseLockSession class from status_manager to repositories
